refactor(pulumi): move module-specific logic to `module.ts`
To keep the same file structure as in terraform plugin.
