+++ plugins/pulumi/index.ts
@@ -8,24 +8,18 @@
 
 import { createGardenPlugin } from "@garden-io/sdk"
 import { dedent } from "@garden-io/sdk/util/string"
-import { configurePulumiModule, deletePulumiDeploy, deployPulumi, getPulumiDeployStatus } from "./handlers"
+import { deletePulumiDeploy, deployPulumi, getPulumiDeployStatus } from "./handlers"
 import { getPulumiCommands } from "./commands"
-
 import { joiVariables } from "@garden-io/core/build/src/config/common"
 import { pulumiCliSPecs } from "./cli"
-import {
-  PulumiDeployConfig,
-  pulumiDeploySchema,
-  PulumiModule,
-  pulumiModuleSchema,
-  pulumiProviderConfigSchema,
-} from "./config"
+import { PulumiDeployConfig, pulumiDeploySchema, pulumiProviderConfigSchema } from "./config"
 import { ExecBuildConfig } from "@garden-io/core/build/src/plugins/exec/build"
 import { join } from "path"
 import { pathExists } from "fs-extra"
 import { ConfigurationError } from "@garden-io/sdk/exceptions"
 import { omit } from "lodash"
 import { ConvertModuleParams } from "@garden-io/core/build/src/plugin/handlers/Module/convert"
+import {configurePulumiModule, PulumiModule, pulumiModuleSchema} from "./module"
 
 // Need to make these variables to avoid escaping issues
 const moduleOutputsTemplateString = "${runtime.services.<module-name>.outputs.<key>}"
